minor optimizations to generator
